ISSUE #2958 - small changes to scrollArea to account for differences in usage
