+++ frontend/src/v5/ui/controls/scrollArea/index.ts
@@ -15,36 +15,4 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import Scrollbars from 'react-custom-scrollbars';
-import styled from 'styled-components';
-
-export const ScrollbarWrapper = styled(Scrollbars).attrs(({ theme }) => ({
-	autoHeight: true,
-	renderThumbVertical: ({ style }) => (
-		<div
-			style={{
-				...style,
-				backgroundColor: theme.palette.base.lightest,
-				right: '6px',
-				bottom: '6px',
-				top: '6px',
-				borderRadius: '3px',
-				width: '6px',
-			}}
-		/>
-	),
-	renderThumbHorizontal: ({ style }) => (
-		<div
-			style={{
-				...style,
-				backgroundColor: theme.palette.base.lightest,
-				left: '6px',
-				right: '6px',
-				bottom: '6px',
-				borderRadius: '3px',
-				height: '6px',
-			}}
-		/>
-	),
-}))`
-`;
+export { ScrollArea } from './scrollArea.component';
